[1.0][sfSympalThemePlugin] Moving some existing app.yml values and adding others.
A few of the values were moved under the "theme" category for consistency.

Two additional sections were added to allow for setting of a theme for modules or routes. This was possible previously, but the module and route key was set at the base of sympal_config which seemed like a bad idea (cluttering and possibly conflicting on that namespace level).