WIP - Refactoring the plugin to be independent of sympal.
I'm also excercising a newer philosophy of removing all sfConfig from service classes. I believe these variables should be passed in as dependencies and so I'm leaning much more on the plugin configuration class to be the context-managing class that creates these service objects and passes in the sfConfig configuration variables as dependencies.

This creates a well-defined context - you have only one class (the configuration class) which impresses the context's static configuration.

This also means that your service classes are now quite independent and never reach statically or globally for any variables.