Removing the sympal_themes module.
Even though the preview function is very cool, it doesn't make sense to allow the user to preview a theme if they can't actually modify it.

For the time being, this has been moved to sfSympalPlugin. Perhaps I'll add an optional dependency on something later (app.yml write) and put the module back in with that dependency.